Print milliseconds when uptime < 60s
It's turning out to be really convenient to get a boot time estimate by
looking at the uptime printed by NervesMOTD on the serial console.
Seconds are too coarse for measuring some optimizations. This adds
milliseconds to the print, but only if the uptime is less than a minute.
This doesn't add milliseconds after that to avoid making the line too
long.