Add serial number
For the longest time, the serial number was super easy to grab from the
hostname. However, the hostname lowercases the serial number so it gets
annoying to manually recapitalize the serial when pasting in a text
sensitive field. This adds the serial number to the default output
unchanged to avoid any issues. Given how useful the serial number, it
seems like it should be listed early.