Use application environment for default options
This makes it possible to just run `NervesMOTD.print` to get the
complete customized version of the MOTD. This is nice when there's
something that NervesMOTD shows and you just want to see it again
without reconnecting.

Any option that can be given to `NervesMOTD.print` can be specified in
the Application environment for `:nerves_motd`. Options passed to
`print/1` take precendence like you'd expect.