Match :inet.gethostname explicitly just in case
If this ever were to fail, the way it does is to give an invalid ANSI
escape sequence error due to the error reason not being a known atom.
This has never happened, but when reviewing the code for invalid ANSI
escape sequence errors, this was one place that was found.