Update uname line to include firmware nickname
This removes firmware architecture and platform since they were making
for a long and awkward line. Will try to add back in a subsequent PR.

This includes fwup's word34567 implementation since it doesn't exist in
another Elixir library yet. Despite the amount of code, it's only a
2.7KB BEAM file when stripped. I expect that when this gets a good home
that it will move out of NervesMOTD.